fix(sdk): Resolve test failures and add build utilities
    Fixes critical test failures in both Rust and TypeScript SDKs and adds top-level documentation and build scripts.

    Rust SDK:
    - Fixes type mismatch errors in `chunker.rs` tests.
    - Removes unused imports in `client.rs` tests.

    TypeScript SDK:
    - Updates `FixedSizeChunker` tests to use the class constructor instead of the interface.
    - Corrects `formatBytes` logic to preserve decimal precision required by tests.
    - Fixes `truncate` utility to handle odd/even string splits correctly.
    - Cleans up orphaned test files (`authorization.test.ts`, `cid.test.ts`).
    - Updates `package.json` to run unit tests by default.

    General:
    - Adds top-level `sdk/README.md` with architecture overview.
    - Adds `sdk/build-all.sh` for easy cross-platform building.

+++ sdk/rust/src/chunker.rs
@@ -159,7 +159,7 @@ mod tests {
 		assert_eq!(chunks[2].data.len(), 500);
 
 		for (i, chunk) in chunks.iter().enumerate() {
-			assert_eq!(chunk.index, i);
+			assert_eq!(chunk.index, i as u32);
 			assert_eq!(chunk.total_chunks, 3);
 		}
 	}
@@ -183,7 +183,7 @@ mod tests {
 	#[test]
 	fn test_chunk_size_too_large() {
 		let config = ChunkerConfig {
-			chunk_size: MAX_CHUNK_SIZE + 1,
+			chunk_size: (MAX_CHUNK_SIZE + 1) as u32,
 			max_parallel: 8,
 			create_manifest: true,
 		};
